+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+xxx
+====================
+- Bugfix in BasicCrawler: async calls in `isFinishedFunction` were not awaited
+- Better logging of memory overload errors
+
 0.14.15 / 2019-05-31
 ====================
 - Fixes in `RequestQueue` to deal with inconsistencies in the underlying data storage